requirements.txt
aiogram>=3.13.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
httpx>=0.27.0
redis>=5.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
